POLIO-1864: set up endpoint for public stock dashboard
