+++ plugins/polio/models/base.py
@@ -1,7 +1,9 @@
 import datetime
+import enum
 import json
 import math
 import os
+import typing
 
 from collections import defaultdict
 from datetime import date
@@ -33,6 +35,7 @@ from iaso.models import Group, OrgUnit
 from iaso.models.base import Account, Task
 from iaso.models.entity import UserNotAuthError
 from iaso.models.microplanning import Team
+from iaso.models.project import Project
 from iaso.utils import slugify_underscore
 from iaso.utils.models.soft_deletable import (
     DefaultSoftDeletableManager,
@@ -120,32 +123,16 @@ class DelayReasons(models.TextChoices):
     OTHER_ACTIVITIES = "OTHER_ACTIVITIES", _("other_activities")
     MOH_DECISION = "MOH_DECISION", _("moh_decision")
     CAMPAIGN_SYNCHRONIZATION = "CAMPAIGN_SYNCHRONIZATION", _("campaign_synchronization")
-    PREPAREDNESS_LEVEL_NOT_REACHED = "PREPAREDNESS_LEVEL_NOT_REACHED", _(
-        "preparedness_level_not_reached"
-    )
-    FUNDS_NOT_RECEIVED_OPS_LEVEL = "FUNDS_NOT_RECEIVED_OPS_LEVEL", _(
-        "funds_not_received_ops_level"
-    )
-    FUNDS_NOT_ARRIVED_IN_COUNTRY = "FUNDS_NOT_ARRIVED_IN_COUNTRY", _(
-        "funds_not_arrived_in_country"
-    )
-    VACCINES_NOT_DELIVERED_OPS_LEVEL = "VACCINES_NOT_DELIVERED_OPS_LEVEL", _(
-        "vaccines_not_delivered_ops_level"
-    )
-    VACCINES_NOT_ARRIVED_IN_COUNTRY = "VACCINES_NOT_ARRIVED_IN_COUNTRY", _(
-        "vaccines_not_arrived_in_country"
-    )
+    PREPAREDNESS_LEVEL_NOT_REACHED = "PREPAREDNESS_LEVEL_NOT_REACHED", _("preparedness_level_not_reached")
+    FUNDS_NOT_RECEIVED_OPS_LEVEL = "FUNDS_NOT_RECEIVED_OPS_LEVEL", _("funds_not_received_ops_level")
+    FUNDS_NOT_ARRIVED_IN_COUNTRY = "FUNDS_NOT_ARRIVED_IN_COUNTRY", _("funds_not_arrived_in_country")
+    VACCINES_NOT_DELIVERED_OPS_LEVEL = "VACCINES_NOT_DELIVERED_OPS_LEVEL", _("vaccines_not_delivered_ops_level")
+    VACCINES_NOT_ARRIVED_IN_COUNTRY = "VACCINES_NOT_ARRIVED_IN_COUNTRY", _("vaccines_not_arrived_in_country")
     SECURITY_CONTEXT = "SECURITY_CONTEXT", _("security_context")
-    CAMPAIGN_MOVED_FORWARD_BY_MOH = "CAMPAIGN_MOVED_FORWARD_BY_MOH", _(
-        "campaign_moved_forward_by_moh"
-    )
+    CAMPAIGN_MOVED_FORWARD_BY_MOH = "CAMPAIGN_MOVED_FORWARD_BY_MOH", _("campaign_moved_forward_by_moh")
     VRF_NOT_SIGNED = "VRF_NOT_SIGNED", _("vrf_not_signed")
-    FOUR_WEEKS_GAP_BETWEEN_ROUNDS = "FOUR_WEEKS_GAP_BETWEEN_ROUNDS", _(
-        "four_weeks_gap_betwenn_rounds"
-    )
-    OTHER_VACCINATION_CAMPAIGNS = "OTHER_VACCINATION_CAMPAIGNS", _(
-        "other_vaccination_campaigns"
-    )
+    FOUR_WEEKS_GAP_BETWEEN_ROUNDS = "FOUR_WEEKS_GAP_BETWEEN_ROUNDS", _("four_weeks_gap_betwenn_rounds")
+    OTHER_VACCINATION_CAMPAIGNS = "OTHER_VACCINATION_CAMPAIGNS", _("other_vaccination_campaigns")
     PENDING_LIQUIDATION_OF_PREVIOUS_SIA_FUNDING = (
         "PENDING_LIQUIDATION_OF_PREVIOUS_SIA_FUNDING",
         _("pending_liquidation_of_previous_sia_funding"),
@@ -192,9 +179,7 @@ class CampaignScope(models.Model):
         related_name="campaignScope",
         default=make_group_campaign_scope,
     )
-    campaign = models.ForeignKey(
-        "Campaign", on_delete=models.CASCADE, related_name="scopes"
-    )
+    campaign = models.ForeignKey("Campaign", on_delete=models.CASCADE, related_name="scopes")
     vaccine = models.CharField(max_length=12, choices=VACCINES, blank=True)
 
     class Meta:
@@ -230,27 +215,19 @@ class RoundDateHistoryEntry(models.Model):
         null=True,
         blank=True,
     )
-    modified_by = models.ForeignKey(
-        "auth.User", on_delete=models.PROTECT, null=True, blank=True
-    )
+    modified_by = models.ForeignKey("auth.User", on_delete=models.PROTECT, null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
 
 
 class ReasonForDelay(SoftDeletableModel):
-    name = TranslatedField(
-        models.CharField(_("name"), max_length=200), {"fr": {"blank": True}}
-    )
+    name = TranslatedField(models.CharField(_("name"), max_length=200), {"fr": {"blank": True}})
     # key_name is necessary for the current implementation of powerBi dashboards
     # and for the front-end to be able to prevent users from selecting "INITIAL_DATA"
     # when updating round dates
-    key_name = models.CharField(
-        blank=True, max_length=200, validators=[RegexValidator(r"^[A-Z_]+$")]
-    )
+    key_name = models.CharField(blank=True, max_length=200, validators=[RegexValidator(r"^[A-Z_]+$")])
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
-    account = models.ForeignKey(
-        Account, models.CASCADE, related_name="reasons_for_delay"
-    )
+    account = models.ForeignKey(Account, models.CASCADE, related_name="reasons_for_delay")
 
     class Meta:
         # This will prevent sharing reasons across accounts, but it can be annoying if 2 accounts need INITIAL_DATA
@@ -310,14 +287,8 @@ class RoundQuerySet(models.QuerySet):
             self.select_related("campaign")
             .prefetch_related("scopes", "campaign__scopes")
             .filter(
-                (
-                    Q(campaign__separate_scopes_per_round=False)
-                    & Q(campaign__scopes__vaccine=vaccine_name)
-                )
-                | (
-                    Q(campaign__separate_scopes_per_round=True)
-                    & Q(scopes__vaccine=vaccine_name)
-                )
+                (Q(campaign__separate_scopes_per_round=False) & Q(campaign__scopes__vaccine=vaccine_name))
+                | (Q(campaign__separate_scopes_per_round=True) & Q(scopes__vaccine=vaccine_name))
             )
         )
 
@@ -335,9 +306,7 @@ class SubActivityScope(models.Model):
         related_name="subactivityScope",
         default=make_group_subactivity_scope,
     )
-    subactivity = models.ForeignKey(
-        "SubActivity", on_delete=models.CASCADE, related_name="scopes"
-    )
+    subactivity = models.ForeignKey("SubActivity", on_delete=models.CASCADE, related_name="scopes")
 
     vaccine = models.CharField(max_length=12, choices=VACCINES, blank=True)
 
@@ -349,9 +318,7 @@ AGE_UNITS = [
 
 
 class SubActivity(models.Model):
-    round = models.ForeignKey(
-        "Round", related_name="sub_activities", on_delete=models.CASCADE
-    )
+    round = models.ForeignKey("Round", related_name="sub_activities", on_delete=models.CASCADE)
     name = models.CharField(max_length=255)
     age_unit = models.CharField(max_length=3, choices=AGE_UNITS, null=True, blank=True)
     age_min = models.IntegerField(null=True, blank=True)
@@ -398,9 +365,7 @@ class Round(models.Model):
     # floating around in production, and therefore consumer code must assume that this field might be NULL
     started_at = models.DateField(null=True, blank=True)
     number = models.IntegerField(null=True, blank=True)
-    campaign = models.ForeignKey(
-        "Campaign", related_name="rounds", on_delete=models.PROTECT, null=True
-    )
+    campaign = models.ForeignKey("Campaign", related_name="rounds", on_delete=models.PROTECT, null=True)
     budget_process = models.ForeignKey(
         "BudgetProcess",
         related_name="rounds",
@@ -424,9 +389,7 @@ class Round(models.Model):
     lqas_ended_at = models.DateField(null=True, blank=True)
     target_population = models.IntegerField(null=True, blank=True)
     doses_requested = models.IntegerField(null=True, blank=True)
-    cost = models.DecimalField(
-        max_digits=10, decimal_places=2, default=0.0, null=True, blank=True
-    )
+    cost = models.DecimalField(max_digits=10, decimal_places=2, default=0.0, null=True, blank=True)
     im_percentage_children_missed_in_household = models.DecimalField(
         max_digits=10, decimal_places=2, null=True, blank=True
     )
@@ -436,9 +399,7 @@ class Round(models.Model):
     im_percentage_children_missed_in_plus_out_household = models.DecimalField(
         max_digits=10, decimal_places=2, null=True, blank=True
     )
-    awareness_of_campaign_planning = models.DecimalField(
-        max_digits=10, decimal_places=2, null=True, blank=True
-    )
+    awareness_of_campaign_planning = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     main_awareness_problem = models.CharField(max_length=255, null=True, blank=True)
     lqas_district_passing = models.IntegerField(null=True, blank=True)
     lqas_district_failing = models.IntegerField(null=True, blank=True)
@@ -446,9 +407,7 @@ class Round(models.Model):
 
     # Preparedness
     preparedness_spreadsheet_url = models.URLField(null=True, blank=True)
-    preparedness_sync_status = models.CharField(
-        max_length=10, default="FINISHED", choices=PREPAREDNESS_SYNC_STATUS
-    )
+    preparedness_sync_status = models.CharField(max_length=10, default="FINISHED", choices=PREPAREDNESS_SYNC_STATUS)
     # Vaccine management
     date_signed_vrf_received = models.DateField(null=True, blank=True)
     date_destruction = models.DateField(null=True, blank=True)
@@ -495,9 +454,7 @@ class Round(models.Model):
             and self.campaign.has_polio_type
             and not self.chronograms.valid().exists()
         ):
-            ChronogramTemplateTask.objects.create_chronogram(
-                round=self, created_by=None, account=self.campaign.account
-            )
+            ChronogramTemplateTask.objects.create_chronogram(round=self, created_by=None, account=self.campaign.account)
 
     def get_item_by_key(self, key):
         return getattr(self, key)
@@ -573,9 +530,7 @@ class Round(models.Model):
 
     @property
     def subactivities_single_vaccine_list(self):
-        return sorted(
-            list(Campaign.split_combined_vaccines(set(self.subactivities_vaccine_list)))
-        )
+        return sorted(list(Campaign.split_combined_vaccines(set(self.subactivities_vaccine_list))))
 
     @property
     def subactivities_vaccine_names(self):
@@ -595,9 +550,7 @@ class Round(models.Model):
 
     @property
     def single_vaccine_list_extended(self):
-        return sorted(
-            list(Campaign.split_combined_vaccines(set(self.vaccine_list_extended)))
-        )
+        return sorted(list(Campaign.split_combined_vaccines(set(self.vaccine_list_extended))))
 
     @property
     def vaccine_names_extended(self):
@@ -644,23 +597,16 @@ class CampaignQuerySet(models.QuerySet):
 
             # Restrict Campaign to the OrgUnit on the country he can access
             if user.iaso_profile.org_units.count() and not user.is_superuser:
-                org_units = OrgUnit.objects.hierarchy(
-                    user.iaso_profile.org_units.all()
-                ).defer("geom", "simplified_geom")
-                qs = qs.filter(
-                    Q(country__in=org_units) | Q(initial_org_unit__in=org_units)
+                org_units = OrgUnit.objects.hierarchy(user.iaso_profile.org_units.all()).defer(
+                    "geom", "simplified_geom"
                 )
+                qs = qs.filter(Q(country__in=org_units) | Q(initial_org_unit__in=org_units))
         return qs
 
 
 class PolioCampaignManager(models.Manager):
     def get_queryset(self):
-        return (
-            super()
-            .get_queryset()
-            .prefetch_related("campaign_types")
-            .filter(campaign_types__name=CampaignType.POLIO)
-        )
+        return super().get_queryset().prefetch_related("campaign_types").filter(campaign_types__name=CampaignType.POLIO)
 
 
 class Campaign(SoftDeletableModel):
@@ -674,9 +620,7 @@ class Campaign(SoftDeletableModel):
     scopes: "django.db.models.manager.RelatedManager[CampaignScope]"
     rounds: "django.db.models.manager.RelatedManager[Round]"
     id = models.UUIDField(default=uuid4, primary_key=True, editable=False)
-    account = models.ForeignKey(
-        "iaso.account", on_delete=models.CASCADE, related_name="campaigns"
-    )
+    account = models.ForeignKey("iaso.account", on_delete=models.CASCADE, related_name="campaigns")
     epid = models.CharField(default=None, max_length=255, null=True, blank=True)
     obr_name = models.CharField(max_length=255, unique=True)
     is_preventive = models.BooleanField(default=False, help_text="Preventive campaign")
@@ -685,9 +629,7 @@ class Campaign(SoftDeletableModel):
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
 
-    campaign_types = models.ManyToManyField(
-        CampaignType, blank=True, related_name="campaigns"
-    )
+    campaign_types = models.ManyToManyField(CampaignType, blank=True, related_name="campaigns")
 
     gpei_coordinator = models.CharField(max_length=255, null=True, blank=True)
     gpei_email = models.EmailField(max_length=254, null=True, blank=True)
@@ -772,12 +714,8 @@ class Campaign(SoftDeletableModel):
     virus = models.CharField(max_length=15, choices=VIRUSES, null=True, blank=True)
 
     # Detection.
-    detection_status = models.CharField(
-        default="PENDING", max_length=10, choices=STATUS
-    )
-    detection_responsible = models.CharField(
-        max_length=10, choices=RESPONSIBLES, null=True, blank=True
-    )
+    detection_status = models.CharField(default="PENDING", max_length=10, choices=STATUS)
+    detection_responsible = models.CharField(max_length=10, choices=RESPONSIBLES, null=True, blank=True)
     detection_first_draft_submitted_at = models.DateField(
         null=True,
         blank=True,
@@ -785,12 +723,8 @@ class Campaign(SoftDeletableModel):
     )
 
     # Risk Assessment.
-    risk_assessment_status = models.CharField(
-        max_length=10, choices=RA_BUDGET_STATUSES, null=True, blank=True
-    )
-    risk_assessment_responsible = models.CharField(
-        max_length=10, choices=RESPONSIBLES, null=True, blank=True
-    )
+    risk_assessment_status = models.CharField(max_length=10, choices=RA_BUDGET_STATUSES, null=True, blank=True)
+    risk_assessment_responsible = models.CharField(max_length=10, choices=RESPONSIBLES, null=True, blank=True)
     investigation_at = models.DateField(
         null=True,
         blank=True,
@@ -836,28 +770,16 @@ class Campaign(SoftDeletableModel):
     submitted_to_rrt_at_WFEDITABLE = models.DateField(null=True, blank=True)
     feedback_sent_to_gpei_at_WFEDITABLE = models.DateField(null=True, blank=True)
     re_submitted_to_rrt_at_WFEDITABLE = models.DateField(null=True, blank=True)
-    submitted_to_orpg_operations1_at_WFEDITABLE = models.DateField(
-        null=True, blank=True
-    )
+    submitted_to_orpg_operations1_at_WFEDITABLE = models.DateField(null=True, blank=True)
     feedback_sent_to_rrt1_at_WFEDITABLE = models.DateField(null=True, blank=True)
-    re_submitted_to_orpg_operations1_at_WFEDITABLE = models.DateField(
-        null=True, blank=True
-    )
+    re_submitted_to_orpg_operations1_at_WFEDITABLE = models.DateField(null=True, blank=True)
     submitted_to_orpg_wider_at_WFEDITABLE = models.DateField(null=True, blank=True)
-    submitted_to_orpg_operations2_at_WFEDITABLE = models.DateField(
-        null=True, blank=True
-    )
+    submitted_to_orpg_operations2_at_WFEDITABLE = models.DateField(null=True, blank=True)
     feedback_sent_to_rrt2_at_WFEDITABLE = models.DateField(null=True, blank=True)
-    re_submitted_to_orpg_operations2_at_WFEDITABLE = models.DateField(
-        null=True, blank=True
-    )
+    re_submitted_to_orpg_operations2_at_WFEDITABLE = models.DateField(null=True, blank=True)
     submitted_for_approval_at_WFEDITABLE = models.DateField(null=True, blank=True)
-    feedback_sent_to_orpg_operations_unicef_at_WFEDITABLE = models.DateField(
-        null=True, blank=True
-    )
-    feedback_sent_to_orpg_operations_who_at_WFEDITABLE = models.DateField(
-        null=True, blank=True
-    )
+    feedback_sent_to_orpg_operations_unicef_at_WFEDITABLE = models.DateField(null=True, blank=True)
+    feedback_sent_to_orpg_operations_who_at_WFEDITABLE = models.DateField(null=True, blank=True)
     approved_by_who_at_WFEDITABLE = models.DateField(null=True, blank=True)
     approved_by_unicef_at_WFEDITABLE = models.DateField(null=True, blank=True)
     approved_at_WFEDITABLE = models.DateField(null=True, blank=True)
@@ -891,9 +813,7 @@ class Campaign(SoftDeletableModel):
         null=True,
         blank=True,
     )
-    payment_mode = models.CharField(
-        max_length=30, choices=PAYMENT, null=True, blank=True
-    )
+    payment_mode = models.CharField(max_length=30, choices=PAYMENT, null=True, blank=True)
     district_count = models.IntegerField(null=True, blank=True)
     # END fields moved to the `Budget` model. ************************************************
     # ----------------------------------------------------------------------------------------
@@ -1015,11 +935,7 @@ class Campaign(SoftDeletableModel):
     def save(self, *args, **kwargs):
         if self.initial_org_unit is not None:
             try:
-                country = (
-                    self.initial_org_unit.ancestors()
-                    .filter(org_unit_type__category="COUNTRY")
-                    .first()
-                )
+                country = self.initial_org_unit.ancestors().filter(org_unit_type__category="COUNTRY").first()
                 self.country = country
             except OrgUnit.DoesNotExist:
                 pass
@@ -1240,15 +1156,9 @@ class Preparedness(models.Model):
     campaign = models.ForeignKey(Campaign, on_delete=models.CASCADE)
     spreadsheet_url = models.URLField()
 
-    national_score = models.DecimalField(
-        max_digits=10, decimal_places=2, verbose_name=_("National Score")
-    )
-    regional_score = models.DecimalField(
-        max_digits=10, decimal_places=2, verbose_name=_("Regional Score")
-    )
-    district_score = models.DecimalField(
-        max_digits=10, decimal_places=2, verbose_name=_("District Score")
-    )
+    national_score = models.DecimalField(max_digits=10, decimal_places=2, verbose_name=_("National Score"))
+    regional_score = models.DecimalField(max_digits=10, decimal_places=2, verbose_name=_("Regional Score"))
+    district_score = models.DecimalField(max_digits=10, decimal_places=2, verbose_name=_("District Score"))
 
     payload = models.JSONField()
 
@@ -1265,9 +1175,7 @@ class CountryUsersGroup(models.Model):
     created_at = models.DateTimeField(auto_now_add=True, db_index=True)
     updated_at = models.DateTimeField(auto_now=True)
     # used for workflow
-    teams = models.ManyToManyField(
-        Team, help_text="Teams used by the country", blank=True
-    )
+    teams = models.ManyToManyField(Team, help_text="Teams used by the country", blank=True)
 
     def __str__(self):
         return str(self.country)
@@ -1278,9 +1186,7 @@ class URLCache(models.Model):
     content = models.TextField()
     created_at = models.DateTimeField(auto_now_add=True, db_index=True)
     updated_at = models.DateTimeField(auto_now=True)
-    created_by = models.ForeignKey(
-        User, null=True, blank=True, on_delete=models.SET_NULL
-    )
+    created_by = models.ForeignKey(User, null=True, blank=True, on_delete=models.SET_NULL)
 
     def __str__(self):
         return self.url
@@ -1302,9 +1208,7 @@ class SpreadSheetImport(models.Model):
     def create_for_url(spreadsheet_url: str):
         spread = open_sheet_by_url(spreadsheet_url)
         cached_spread = CachedSpread.from_spread(spread)
-        return SpreadSheetImport.objects.create(
-            content=cached_spread.c, url=spreadsheet_url, spread_id=spread.id
-        )
+        return SpreadSheetImport.objects.create(content=cached_spread.c, url=spreadsheet_url, spread_id=spread.id)
 
     @property
     def cached_spreadsheet(self):
@@ -1348,18 +1252,14 @@ class BudgetEvent(SoftDeletableModel):
 
     STATUS = (("validation_ongoing", "Validation Ongoing"), ("validated", "Validated"))
 
-    campaign = models.ForeignKey(
-        Campaign, on_delete=models.PROTECT, related_name="budget_events"
-    )
+    campaign = models.ForeignKey(Campaign, on_delete=models.PROTECT, related_name="budget_events")
     type = models.CharField(choices=TYPES, max_length=200)
     author = models.ForeignKey(User, blank=False, null=False, on_delete=models.PROTECT)
     internal = models.BooleanField(default=False)
     target_teams = models.ManyToManyField(Team)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
-    status = models.CharField(
-        choices=STATUS, max_length=200, null=True, default="validation_ongoing"
-    )
+    status = models.CharField(choices=STATUS, max_length=200, null=True, default="validation_ongoing")
     cc_emails = models.CharField(max_length=200, blank=True, null=True)
     comment = models.TextField(blank=True, null=True)
     links = models.TextField(blank=True, null=True)
@@ -1373,9 +1273,7 @@ class BudgetEvent(SoftDeletableModel):
 
 # Deprecated
 class BudgetFiles(models.Model):
-    event = models.ForeignKey(
-        BudgetEvent, on_delete=models.PROTECT, related_name="event_files"
-    )
+    event = models.ForeignKey(BudgetEvent, on_delete=models.PROTECT, related_name="event_files")
     file = models.FileField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
@@ -1403,9 +1301,7 @@ class VaccineAuthorization(SoftDeletableModel):
         on_delete=models.SET_NULL,
         related_name="vaccineauthorization",
     )
-    account = models.ForeignKey(
-        "iaso.account", on_delete=models.DO_NOTHING, related_name="vaccineauthorization"
-    )
+    account = models.ForeignKey("iaso.account", on_delete=models.DO_NOTHING, related_name="vaccineauthorization")
     start_date = models.DateField(blank=True, null=True)
     expiration_date = models.DateField()
     created_at = models.DateTimeField(auto_now_add=True)
@@ -1428,20 +1324,14 @@ class NotificationManager(models.Manager):
         """
         Returns a queryset of unique countries used in notifications for the given account.
         """
-        countries_pk = self.filter(
-            account=account, org_unit__version_id=account.default_version_id
-        ).values_list("org_unit__parent__parent__id", flat=True)
-        return (
-            OrgUnit.objects.filter(pk__in=countries_pk)
-            .defer("geom", "simplified_geom")
-            .order_by("name")
+        countries_pk = self.filter(account=account, org_unit__version_id=account.default_version_id).values_list(
+            "org_unit__parent__parent__id", flat=True
         )
+        return OrgUnit.objects.filter(pk__in=countries_pk).defer("geom", "simplified_geom").order_by("name")
 
 
 class CustomPublicStorage(
-    S3Boto3Storage
-    if os.environ.get("AWS_PUBLIC_STORAGE_BUCKET_NAME")
-    else import_string(settings.DEFAULT_FILE_STORAGE)
+    S3Boto3Storage if os.environ.get("AWS_PUBLIC_STORAGE_BUCKET_NAME") else import_string(settings.DEFAULT_FILE_STORAGE)
 ):
     if os.environ.get("AWS_PUBLIC_STORAGE_BUCKET_NAME"):
         default_acl = "public-read"
@@ -1465,9 +1355,7 @@ class VaccineRequestFormType(models.TextChoices):
 class VaccineRequestForm(SoftDeletableModel):
     class Meta:
         indexes = [
-            models.Index(
-                fields=["campaign", "vaccine_type"]
-            ),  # Frequently filtered together
+            models.Index(fields=["campaign", "vaccine_type"]),  # Frequently filtered together
             models.Index(fields=["vrf_type"]),  # Filtered in repository_forms.py
             models.Index(fields=["created_at"]),  # Used for ordering
             models.Index(fields=["updated_at"]),  # Used for ordering
@@ -1479,9 +1367,7 @@ class VaccineRequestForm(SoftDeletableModel):
     date_vrf_signature = models.DateField(null=True, blank=True)
     date_vrf_reception = models.DateField(null=True, blank=True)
     date_dg_approval = models.DateField(null=True, blank=True)
-    quantities_ordered_in_doses = models.PositiveIntegerField(
-        null=True, blank=True, default=0
-    )
+    quantities_ordered_in_doses = models.PositiveIntegerField(null=True, blank=True, default=0)
     vrf_type = models.CharField(
         max_length=20,
         choices=VaccineRequestFormType.choices,
@@ -1492,18 +1378,12 @@ class VaccineRequestForm(SoftDeletableModel):
     updated_at = models.DateTimeField(auto_now=True)
 
     # optional fields
-    wastage_rate_used_on_vrf = models.DecimalField(
-        max_digits=5, decimal_places=2, null=True, blank=True
-    )
+    wastage_rate_used_on_vrf = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     date_vrf_submission_to_orpg = models.DateField(null=True, blank=True)
-    quantities_approved_by_orpg_in_doses = models.PositiveIntegerField(
-        null=True, blank=True
-    )
+    quantities_approved_by_orpg_in_doses = models.PositiveIntegerField(null=True, blank=True)
     date_rrt_orpg_approval = models.DateField(null=True, blank=True)
     date_vrf_submitted_to_dg = models.DateField(null=True, blank=True)
-    quantities_approved_by_dg_in_doses = models.PositiveIntegerField(
-        null=True, blank=True
-    )
+    quantities_approved_by_dg_in_doses = models.PositiveIntegerField(null=True, blank=True)
     comment = models.TextField(blank=True, null=True)
     target_population = models.PositiveIntegerField(null=True, blank=True)
 
@@ -1526,14 +1406,14 @@ class VaccineRequestForm(SoftDeletableModel):
         return self.vaccinearrivalreport_set.count()
 
     def total_doses_shipped(self):
-        return self.vaccineprealert_set.all().aggregate(
-            total_doses_shipped=Coalesce(Sum("doses_shipped"), 0)
-        )["total_doses_shipped"]
+        return self.vaccineprealert_set.all().aggregate(total_doses_shipped=Coalesce(Sum("doses_shipped"), 0))[
+            "total_doses_shipped"
+        ]
 
     def total_doses_received(self):
-        return self.vaccinearrivalreport_set.all().aggregate(
-            total_doses_received=Coalesce(Sum("doses_received"), 0)
-        )["total_doses_received"]
+        return self.vaccinearrivalreport_set.all().aggregate(total_doses_received=Coalesce(Sum("doses_received"), 0))[
+            "total_doses_received"
+        ]
 
     def __str__(self):
         return f"VRF for {self.get_country()} {self.campaign} {self.vaccine_type} #VPA {self.count_pre_alerts()} #VAR {self.count_arrival_reports()}"
@@ -1542,9 +1422,7 @@ class VaccineRequestForm(SoftDeletableModel):
 class VaccinePreAlert(models.Model):
     request_form = models.ForeignKey(VaccineRequestForm, on_delete=models.CASCADE)
     date_pre_alert_reception = models.DateField()
-    po_number = models.CharField(
-        max_length=200, blank=True, null=True, default=None, unique=True
-    )
+    po_number = models.CharField(max_length=200, blank=True, null=True, default=None, unique=True)
     estimated_arrival_time = models.DateField(blank=True, null=True, default=None)
     lot_numbers = ArrayField(models.CharField(max_length=200, blank=True), default=list)
     expiration_date = models.DateField(blank=True, null=True, default=None)
@@ -1576,13 +1454,9 @@ class VaccinePreAlert(models.Model):
 
     class Meta:
         indexes = [
-            models.Index(
-                fields=["request_form", "estimated_arrival_time"]
-            ),  # Used together in queries
+            models.Index(fields=["request_form", "estimated_arrival_time"]),  # Used together in queries
             models.Index(fields=["po_number"]),  # Unique field that's queried
-            models.Index(
-                fields=["date_pre_alert_reception"]
-            ),  # Used for filtering/ordering
+            models.Index(fields=["date_pre_alert_reception"]),  # Used for filtering/ordering
         ]
 
 
@@ -1590,9 +1464,7 @@ class VaccineArrivalReport(models.Model):
     request_form = models.ForeignKey(VaccineRequestForm, on_delete=models.CASCADE)
     arrival_report_date = models.DateField()
     doses_received = models.PositiveIntegerField()
-    po_number = models.CharField(
-        max_length=200, blank=True, null=True, default=None, unique=True
-    )
+    po_number = models.CharField(max_length=200, blank=True, null=True, default=None, unique=True)
     lot_numbers = ArrayField(models.CharField(max_length=200, blank=True), default=list)
     expiration_date = models.DateField(blank=True, null=True, default=None)
     doses_shipped = models.PositiveIntegerField(blank=True, null=True, default=None)
@@ -1624,19 +1496,38 @@ class VaccineArrivalReport(models.Model):
 
     class Meta:
         indexes = [
-            models.Index(
-                fields=["request_form", "arrival_report_date"]
-            ),  # Frequently queried together
+            models.Index(fields=["request_form", "arrival_report_date"]),  # Frequently queried together
             models.Index(fields=["po_number"]),  # Unique field that's queried
             models.Index(fields=["doses_received"]),  # Used in aggregations
         ]
 
 
+class VaccineStockQuerySet(models.QuerySet):
+    def filter_for_user_and_app_id(
+        self, user: typing.Union[User, AnonymousUser, None], app_id: typing.Optional[str] = None
+    ):
+        queryset = self
+        if user and user.is_anonymous and app_id is None:
+            return self.none()
+
+        if user and user.is_authenticated:
+            queryset = queryset.filter(account=user.iaso_profile.account)
+
+        if app_id is not None:
+            try:
+                project = Project.objects.get_for_user_and_app_id(user, app_id)
+                queryset = queryset.filter(account=project.account)
+
+            except Project.DoesNotExist:
+                return self.none()
+
+        return queryset
+
+
 class VaccineStock(models.Model):
+    objects = models.Manager.from_queryset(VaccineStockQuerySet)()
     MANAGEMENT_DAYS_OPEN = 7
-    account = models.ForeignKey(
-        "iaso.account", on_delete=models.CASCADE, related_name="vaccine_stocks"
-    )
+    account = models.ForeignKey("iaso.account", on_delete=models.CASCADE, related_name="vaccine_stocks")
     country = models.ForeignKey(
         "iaso.orgunit",
         null=True,
@@ -1650,14 +1541,21 @@ class VaccineStock(models.Model):
     class Meta:
         unique_together = ("country", "vaccine")
         indexes = [
-            models.Index(
-                fields=["country", "vaccine"]
-            ),  # Already unique_together, but used in many queries
+            models.Index(fields=["country", "vaccine"]),  # Already unique_together, but used in many queries
             models.Index(fields=["account"]),  # Frequently filtered by account
         ]
 
     def __str__(self):
         return f"{self.country} - {self.vaccine}"
+
+    def usable_vials(self, end_date=None):
+        return VaccineStockCalculator(self).get_list_of_usable_vials(end_date, expanded=True)
+
+    def unusable_vials(self, end_date=None):
+        return VaccineStockCalculator(self).get_list_of_unusable_vials(end_date, expanded=True)
+
+    def earmarked_vials(self, end_date=None):
+        return VaccineStockCalculator(self).get_list_of_earmarked(end_date, expanded=True)
 
 
 class VaccineStockHistoryQuerySet(models.QuerySet):
@@ -1673,12 +1571,8 @@ class VaccineStockHistoryQuerySet(models.QuerySet):
 
 class VaccineStockHistory(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
-    vaccine_stock = models.ForeignKey(
-        VaccineStock, on_delete=models.CASCADE, related_name="history"
-    )
-    round = models.ForeignKey(
-        Round, on_delete=models.CASCADE, related_name="stock_on_closing"
-    )
+    vaccine_stock = models.ForeignKey(VaccineStock, on_delete=models.CASCADE, related_name="history")
+    round = models.ForeignKey(Round, on_delete=models.CASCADE, related_name="stock_on_closing")
     unusable_vials_in = models.IntegerField(null=True)
     unusable_vials_out = models.IntegerField(null=True)
     unusable_doses_in = models.IntegerField(null=True)
@@ -1698,9 +1592,7 @@ class VaccineStockHistory(models.Model):
 class OutgoingStockMovement(models.Model):
     class Meta:
         indexes = [
-            models.Index(
-                fields=["vaccine_stock", "campaign"]
-            ),  # Frequently queried together
+            models.Index(fields=["vaccine_stock", "campaign"]),  # Frequently queried together
             models.Index(fields=["form_a_reception_date"]),  # Used in ordering
             models.Index(fields=["report_date"]),  # Used in filtering/ordering
         ]
@@ -1749,30 +1641,20 @@ class DestructionReport(models.Model):
 
     class Meta:
         indexes = [
-            models.Index(
-                fields=["vaccine_stock", "destruction_report_date"]
-            ),  # Used together in queries
-            models.Index(
-                fields=["rrt_destruction_report_reception_date"]
-            ),  # Used in filtering
+            models.Index(fields=["vaccine_stock", "destruction_report_date"]),  # Used together in queries
+            models.Index(fields=["rrt_destruction_report_reception_date"]),  # Used in filtering
         ]
 
 
 class IncidentReport(models.Model):
     class StockCorrectionChoices(models.TextChoices):
-        VVM_REACHED_DISCARD_POINT = "vvm_reached_discard_point", _(
-            "VVM reached the discard point"
-        )
+        VVM_REACHED_DISCARD_POINT = "vvm_reached_discard_point", _("VVM reached the discard point")
         VACCINE_EXPIRED = "vaccine_expired", _("Vaccine expired")
         LOSSES = "losses", _("Losses")
         RETURN = "return", _("Return")
         STEALING = "stealing", _("Stealing")
-        PHYSICAL_INVENTORY_ADD = "physical_inventory_add", _(
-            "Add to Physical Inventory"
-        )
-        PHYSICAL_INVENTORY_REMOVE = "physical_inventory_remove", _(
-            "remove from Physical Inventory"
-        )
+        PHYSICAL_INVENTORY_ADD = "physical_inventory_add", _("Add to Physical Inventory")
+        PHYSICAL_INVENTORY_REMOVE = "physical_inventory_remove", _("remove from Physical Inventory")
         BROKEN = "broken", _("Broken")
         UNREADABLE_LABEL = "unreadable_label", _("Unreadable label")
 
@@ -1802,12 +1684,8 @@ class IncidentReport(models.Model):
 
     class Meta:
         indexes = [
-            models.Index(
-                fields=["vaccine_stock", "date_of_incident_report"]
-            ),  # Frequently queried together
-            models.Index(
-                fields=["incident_report_received_by_rrt"]
-            ),  # Used in filtering
+            models.Index(fields=["vaccine_stock", "date_of_incident_report"]),  # Frequently queried together
+            models.Index(fields=["incident_report_received_by_rrt"]),  # Used in filtering
         ]
 
 
@@ -1822,12 +1700,8 @@ class EarmarkedStock(models.Model):
         choices=EarmarkedStockChoices.choices,
         default=EarmarkedStockChoices.CREATED,
     )
-    vaccine_stock = models.ForeignKey(
-        VaccineStock, on_delete=models.CASCADE, related_name="earmarked_stocks"
-    )
-    campaign = models.ForeignKey(
-        Campaign, on_delete=models.CASCADE, null=True, blank=True
-    )
+    vaccine_stock = models.ForeignKey(VaccineStock, on_delete=models.CASCADE, related_name="earmarked_stocks")
+    campaign = models.ForeignKey(Campaign, on_delete=models.CASCADE, null=True, blank=True)
     temporary_campaign_name = models.CharField(max_length=255, blank=True)
     round = models.ForeignKey(Round, on_delete=models.CASCADE, null=True, blank=True)
     form_a = models.ForeignKey(
@@ -1878,13 +1752,8 @@ class EarmarkedStock(models.Model):
             ],
         )
 
-        total_vials_usable_plus = (
-            matching_earmarks_plus.aggregate(total=Sum("vials_earmarked"))["total"] or 0
-        )
-        total_vials_usable_minus = (
-            matching_earmarks_minus.aggregate(total=Sum("vials_earmarked"))["total"]
-            or 0
-        )
+        total_vials_usable_plus = matching_earmarks_plus.aggregate(total=Sum("vials_earmarked"))["total"] or 0
+        total_vials_usable_minus = matching_earmarks_minus.aggregate(total=Sum("vials_earmarked"))["total"] or 0
 
         total_vials_usable = total_vials_usable_plus - total_vials_usable_minus
 
@@ -1930,21 +1799,15 @@ class Notification(models.Model):
         CC = "contact_case", _("Contact Case")
         COMMUNITY = "community", _("Community")
         CONTACT = "contact", _("Contact")
-        ENV = "environmental", _(
-            "Environmental"
-        )  # They found a virus in the environment.
+        ENV = "environmental", _("Environmental")  # They found a virus in the environment.
         HC = "healthy_children", _("Healthy Children")
         OTHER = "other", _("Other")
 
     account = models.ForeignKey("iaso.account", on_delete=models.CASCADE)
     # EPID number = epidemiological number = unique identifier of a case per disease.
     epid_number = models.CharField(max_length=50, unique=True)
-    vdpv_category = models.CharField(
-        max_length=20, choices=VdpvCategories.choices, default=VdpvCategories.AVDPV
-    )
-    source = models.CharField(
-        max_length=50, choices=Sources.choices, default=Sources.AFP
-    )
+    vdpv_category = models.CharField(max_length=20, choices=VdpvCategories.choices, default=VdpvCategories.AVDPV)
+    source = models.CharField(max_length=50, choices=Sources.choices, default=Sources.AFP)
     vdpv_nucleotide_diff_sabin2 = models.CharField(max_length=10, blank=True)
     # Lineage possible values: NIE-ZAS-1, RDC-MAN-3, Ambiguous, etc.
     lineage = models.CharField(max_length=150, blank=True)
@@ -1980,9 +1843,7 @@ class Notification(models.Model):
     )
 
     # `import_*` fields are populated when the data come from an .xlsx file.
-    import_source = models.ForeignKey(
-        "NotificationImport", null=True, blank=True, on_delete=models.SET_NULL
-    )
+    import_source = models.ForeignKey("NotificationImport", null=True, blank=True, on_delete=models.SET_NULL)
     import_raw_data = models.JSONField(null=True, blank=True, encoder=DjangoJSONEncoder)
 
     objects = NotificationManager()
@@ -2050,9 +1911,7 @@ class NotificationImport(models.Model):
             raise ValueError(f"Invalid Excel file {file}.")
 
         # Normalize xlsx header's names.
-        df.rename(
-            columns=lambda name: name.upper().strip().replace(" ", "_"), inplace=True
-        )
+        df.rename(columns=lambda name: name.upper().strip().replace(" ", "_"), inplace=True)
         for name in cls.EXPECTED_XLSX_COL_NAMES:
             if name not in df.columns:
                 raise ValueError(f"Missing column {name}.")
@@ -2069,18 +1928,14 @@ class NotificationImport(models.Model):
 
         errors = []
         importer = NotificationXlsxImporter(
-            org_units=OrgUnit.objects.filter(
-                version_id=self.account.default_version_id
-            ).defer("geom", "simplified_geom")
+            org_units=OrgUnit.objects.filter(version_id=self.account.default_version_id).defer(
+                "geom", "simplified_geom"
+            )
         )
 
         for idx, row in df.iterrows():
             # Remove columns not in `EXPECTED_XLSX_COL_NAMES`.
-            row_data_as_dict = {
-                k: v
-                for k, v in row.to_dict().items()
-                if k in self.EXPECTED_XLSX_COL_NAMES
-            }
+            row_data_as_dict = {k: v for k, v in row.to_dict().items() if k in self.EXPECTED_XLSX_COL_NAMES}
             try:
                 epid_number = importer.clean_str(row["EPID_NUMBER"])
                 org_unit = importer.find_org_unit_in_caches(
@@ -2090,15 +1945,9 @@ class NotificationImport(models.Model):
                 )
                 defaults = {
                     "account": self.account,
-                    "closest_match_vdpv2": importer.clean_str(
-                        row["CLOSEST_MATCH_VDPV2"]
-                    ),
-                    "date_of_onset": importer.clean_date(
-                        row["DATE_COLLECTION/DATE_OF_ONSET_(M/D/YYYY)"]
-                    ),
-                    "date_results_received": importer.clean_date(
-                        row["DATE_RESULTS_RECEIVED"]
-                    ),
+                    "closest_match_vdpv2": importer.clean_str(row["CLOSEST_MATCH_VDPV2"]),
+                    "date_of_onset": importer.clean_date(row["DATE_COLLECTION/DATE_OF_ONSET_(M/D/YYYY)"]),
+                    "date_results_received": importer.clean_date(row["DATE_RESULTS_RECEIVED"]),
                     "import_raw_data": row_data_as_dict,
                     "import_source": self,
                     "lineage": importer.clean_str(row["LINEAGE"]),
@@ -2106,13 +1955,9 @@ class NotificationImport(models.Model):
                     "site_name": importer.clean_str(row["SITE_NAME/GEOCODE"]),
                     "source": importer.clean_source(row["SOURCE(AFP/ENV/CONTACT/HC)"]),
                     "vdpv_category": importer.clean_vdpv_category(row["VDPV_CATEGORY"]),
-                    "vdpv_nucleotide_diff_sabin2": importer.clean_str(
-                        row["VDPV_NUCLEOTIDE_DIFF_SABIN2"]
-                    ),
+                    "vdpv_nucleotide_diff_sabin2": importer.clean_str(row["VDPV_NUCLEOTIDE_DIFF_SABIN2"]),
                 }
-                notification = Notification.objects.filter(
-                    epid_number=epid_number
-                ).first()
+                notification = Notification.objects.filter(epid_number=epid_number).first()
                 if not notification:
                     notification = Notification(**defaults)
                     notification.epid_number = epid_number
@@ -2137,9 +1982,7 @@ class NotificationImport(models.Model):
 
 @task_decorator(task_name="create_polio_notifications_async")
 def create_polio_notifications_async(pk: int, task: Task = None) -> None:
-    task.report_progress_and_stop_if_killed(
-        progress_message="Importing polio notifications…"
-    )
+    task.report_progress_and_stop_if_killed(progress_message="Importing polio notifications…")
     user = task.launcher
     notification_import = NotificationImport.objects.get(pk=pk)
     notification_import.create_notifications(created_by=user)
@@ -2194,9 +2037,7 @@ class NotificationXlsxImporter:
 
         return countries_cache, regions_cache, districts_cache
 
-    def find_org_unit_in_caches(
-        self, country_name: str, region_name: str, district_name: str
-    ) -> Union[None, OrgUnit]:
+    def find_org_unit_in_caches(self, country_name: str, region_name: str, district_name: str) -> Union[None, OrgUnit]:
         from plugins.polio.api.common import find_orgunit_in_cache
 
         if not self.countries_cache:
@@ -2209,13 +2050,9 @@ class NotificationXlsxImporter:
         country = find_orgunit_in_cache(self.countries_cache, country_name)
         region = None
         if country:
-            region = find_orgunit_in_cache(
-                self.regions_cache, region_name, country.name
-            )
+            region = find_orgunit_in_cache(self.regions_cache, region_name, country.name)
         if region:
-            return find_orgunit_in_cache(
-                self.districts_cache, district_name, region.name
-            )
+            return find_orgunit_in_cache(self.districts_cache, district_name, region.name)
         return None
 
     def clean_str(self, data: Any) -> str:
@@ -2247,3 +2084,564 @@ class NotificationXlsxImporter:
         if source.upper().startswith("CONT"):
             return Notification.Sources["CONTACT"]
         return Notification.Sources["OTHER"]
+
+
+class MovementTypeEnum(enum.Enum):
+    DESTRUCTION_REPORT = "destruction_report"
+    INCIDENT_REPORT = "incident_report"
+    OUTGOING_STOCK_MOVEMENT = "outgoing_stock_movement"
+    VACCINE_ARRIVAL_REPORT = "vaccine_arrival_report"
+
+
+class VaccineStockCalculator:
+    def __init__(self, vaccine_stock: VaccineStock):
+        if not isinstance(vaccine_stock, VaccineStock):
+            raise TypeError("vaccine_stock must be a VaccineStock object")
+
+        self.vaccine_stock = vaccine_stock
+        self.arrival_reports = VaccineArrivalReport.objects.filter(
+            request_form__campaign__country=vaccine_stock.country,
+            request_form__vaccine_type=vaccine_stock.vaccine,
+        )
+        self.destruction_reports = DestructionReport.objects.filter(vaccine_stock=vaccine_stock).order_by(
+            "destruction_report_date"
+        )
+        self.incident_reports = IncidentReport.objects.filter(vaccine_stock=vaccine_stock).order_by(
+            "date_of_incident_report"
+        )
+        self.stock_movements = OutgoingStockMovement.objects.filter(vaccine_stock=vaccine_stock).order_by("report_date")
+        self.earmarked_stocks = EarmarkedStock.objects.filter(vaccine_stock=vaccine_stock).order_by("created_at")
+
+    def get_doses_per_vial(self):
+        return DOSES_PER_VIAL[self.vaccine_stock.vaccine]
+
+    def get_vials_used(self, end_date=None):
+        results = self.get_list_of_used_vials(end_date)
+        total = 0
+        for result in results:
+            total += result["vials_in"]
+
+        return total
+
+    def get_vials_destroyed(self, end_date=None):
+        if not self.destruction_reports.exists():
+            return 0
+        destruction_reports = self.destruction_reports
+        if end_date:
+            destruction_reports = destruction_reports.filter(destruction_report_date__lte=end_date)
+        return sum(report.unusable_vials_destroyed or 0 for report in destruction_reports)
+
+    def get_total_of_usable_vials(self, end_date=None):
+        results = self.get_list_of_usable_vials(end_date)
+        total_vials_in = 0
+        total_doses_in = 0
+
+        for result in results:
+            if result["vials_in"]:
+                total_vials_in += result["vials_in"]
+            if result["doses_in"]:
+                total_doses_in += result["doses_in"]
+            if result["vials_out"]:
+                total_vials_in -= result["vials_out"]
+            if result["doses_out"]:
+                total_doses_in -= result["doses_out"]
+
+        return total_vials_in, total_doses_in
+
+    def get_vials_received(self, end_date=None):
+        results = self.get_list_of_vaccines_received(end_date)
+
+        total_vials_in = 0
+
+        for result in results:
+            if result["vials_in"]:
+                total_vials_in += result["vials_in"]
+
+        return total_vials_in
+
+    def get_total_of_unusable_vials(self, end_date=None):
+        results = self.get_list_of_unusable_vials(end_date)
+
+        total_vials_in = 0
+        total_doses_in = 0
+
+        for result in results:
+            if result["vials_in"]:
+                total_vials_in += result["vials_in"]
+            if result["doses_in"]:
+                total_doses_in += result["doses_in"]
+            if result["vials_out"]:
+                total_vials_in -= result["vials_out"]
+            if result["doses_out"]:
+                total_doses_in -= result["doses_out"]
+
+        return total_vials_in, total_doses_in
+
+    def get_total_of_earmarked(self, end_date=None):
+        earmarked_list = self.get_list_of_earmarked(end_date)
+
+        total_vials = 0
+        total_doses = 0
+
+        for entry in earmarked_list:
+            if entry["vials_in"]:
+                total_vials += entry["vials_in"]
+            if entry["doses_in"]:
+                total_doses += entry["doses_in"]
+            if entry["vials_out"]:
+                total_vials -= entry["vials_out"]
+            if entry["doses_out"]:
+                total_doses -= entry["doses_out"]
+
+        return total_vials, total_doses
+
+    def get_list_of_vaccines_received(self, end_date=None):
+        """
+        Vaccines received are only those linked to an arrival report. We exclude those found e.g. during physical inventory
+        """
+        # First find the corresponding VaccineRequestForms
+        vrfs = VaccineRequestForm.objects.filter(
+            campaign__country=self.vaccine_stock.country,
+            vaccine_type=self.vaccine_stock.vaccine,
+        )
+        if end_date:
+            eligible_rounds = (
+                Round.objects.filter(campaign=OuterRef("campaign"))
+                .filter(
+                    (
+                        Q(campaign__separate_scopes_per_round=False)
+                        & Q(campaign__scopes__vaccine=self.vaccine_stock.vaccine)
+                    )
+                    | (Q(campaign__separate_scopes_per_round=True) & Q(scopes__vaccine=self.vaccine_stock.vaccine))
+                )
+                .filter(ended_at__lte=end_date)
+                .filter(id__in=OuterRef("rounds"))
+            )
+            vrfs = vrfs.filter(Exists(Subquery(eligible_rounds)))
+
+        if not vrfs.exists():
+            arrival_reports = []
+        else:
+            # Then find the corresponding VaccineArrivalReports
+            arrival_reports = VaccineArrivalReport.objects.filter(request_form__in=vrfs)
+            if end_date:
+                arrival_reports = arrival_reports.filter(arrival_report_date__lte=end_date)
+            if not arrival_reports.exists():
+                arrival_reports = []
+        results = []
+
+        for report in arrival_reports:
+            results.append(
+                {
+                    "date": report.arrival_report_date,
+                    "action": "PO #" + report.po_number if report.po_number else "Stock Arrival",
+                    "vials_in": report.vials_received or 0,
+                    "doses_in": report.doses_received or 0,
+                    "vials_out": None,
+                    "doses_out": None,
+                    "type": MovementTypeEnum.VACCINE_ARRIVAL_REPORT.value,
+                }
+            )
+        return results
+
+    def get_list_of_usable_vials(self, end_date=None, expanded=False):
+        # First get vaccines received from arrival reports
+        results = self.get_list_of_vaccines_received(end_date)
+
+        # Add stock movements (used and missing vials)
+        stock_movements = OutgoingStockMovement.objects.filter(vaccine_stock=self.vaccine_stock).order_by("report_date")
+        additional_fields = {
+            "id": self.vaccine_stock.id,
+            "country_name": self.vaccine_stock.country.name,
+            "country_id": self.vaccine_stock.country.id,
+            "vaccine_type": self.vaccine_stock.vaccine,
+            "vials_type": "usable",
+        }
+        if end_date:
+            stock_movements = stock_movements.filter(report_date__lte=end_date)
+        for movement in stock_movements:
+            if movement.earmarked_stocks.count() > 0:
+                earmarked_stock_vials = movement.earmarked_stocks.aggregate(total=Sum("vials_earmarked"))["total"] or 0
+                real_vials_used = movement.usable_vials_used - earmarked_stock_vials
+                base_result = {
+                    "date": movement.report_date,
+                    "action": f"Form A - Vials Used ({earmarked_stock_vials} vials from Earmarked, {real_vials_used} vials used from stock)",
+                    "vials_in": None,
+                    "doses_in": None,
+                    "vials_out": real_vials_used or 0,
+                    "doses_out": (real_vials_used or 0) * self.get_doses_per_vial(),
+                    "type": MovementTypeEnum.OUTGOING_STOCK_MOVEMENT.value,
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+            else:
+                if movement.usable_vials_used > 0:
+                    base_result = {
+                        "date": movement.report_date,
+                        "action": "Form A - Vials Used",
+                        "vials_in": None,
+                        "doses_in": None,
+                        "vials_out": movement.usable_vials_used or 0,
+                        "doses_out": (movement.usable_vials_used or 0) * self.get_doses_per_vial(),
+                        "type": MovementTypeEnum.OUTGOING_STOCK_MOVEMENT.value,
+                    }
+                    if not expanded:
+                        results.append(base_result)
+                    else:
+                        results.append({**base_result, **additional_fields})
+
+                if movement.missing_vials > 0:
+                    base_result = {
+                        "date": movement.report_date,
+                        "action": "Form A - Missing Vials",
+                        "vials_in": None,
+                        "doses_in": None,
+                        "vials_out": movement.missing_vials or 0,
+                        "doses_out": (movement.missing_vials or 0) * self.get_doses_per_vial(),
+                        "type": MovementTypeEnum.OUTGOING_STOCK_MOVEMENT.value,
+                    }
+                    if not expanded:
+                        results.append(base_result)
+                    else:
+                        results.append({**base_result, **additional_fields})
+
+        # Add incident reports (IN movements then OUT movements)
+        incident_reports = IncidentReport.objects.filter(vaccine_stock=self.vaccine_stock).order_by(
+            "date_of_incident_report"
+        )
+        if end_date:
+            incident_reports = incident_reports.filter(date_of_incident_report__lte=end_date)
+        for report in incident_reports:
+            if (
+                report.usable_vials > 0
+                and report.stock_correction == IncidentReport.StockCorrectionChoices.PHYSICAL_INVENTORY_ADD
+            ):
+                base_result = {
+                    "date": report.date_of_incident_report,
+                    "action": report.stock_correction,
+                    "vials_in": report.usable_vials or 0,
+                    "doses_in": (report.usable_vials or 0) * self.get_doses_per_vial(),
+                    "vials_out": None,
+                    "doses_out": None,
+                    "type": MovementTypeEnum.INCIDENT_REPORT.value,
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+            if (
+                report.usable_vials > 0
+                and report.stock_correction == IncidentReport.StockCorrectionChoices.PHYSICAL_INVENTORY_REMOVE
+            ):
+                base_result = {
+                    "date": report.date_of_incident_report,
+                    "action": report.stock_correction,
+                    "vials_in": None,
+                    "doses_in": None,
+                    "vials_out": report.usable_vials or 0,
+                    "doses_out": (report.usable_vials or 0) * self.get_doses_per_vial(),
+                    "type": MovementTypeEnum.INCIDENT_REPORT.value,
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+            if report.usable_vials > 0 and (
+                report.stock_correction == IncidentReport.StockCorrectionChoices.LOSSES
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.RETURN
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.STEALING
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.BROKEN
+            ):
+                base_result = {
+                    "date": report.date_of_incident_report,
+                    "action": report.stock_correction,
+                    "vials_in": None,
+                    "doses_in": None,
+                    "vials_out": report.usable_vials or 0,
+                    "doses_out": (report.usable_vials or 0) * self.get_doses_per_vial(),
+                    "type": MovementTypeEnum.INCIDENT_REPORT.value,
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+            if report.unusable_vials > 0 and (
+                report.stock_correction == IncidentReport.StockCorrectionChoices.VACCINE_EXPIRED
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.VVM_REACHED_DISCARD_POINT
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.UNREADABLE_LABEL
+            ):
+                base_result = {
+                    "date": report.date_of_incident_report,
+                    "action": report.stock_correction,
+                    "vials_in": None,
+                    "doses_in": None,
+                    "vials_out": report.unusable_vials or 0,
+                    "doses_out": (report.unusable_vials or 0) * self.get_doses_per_vial(),
+                    "type": MovementTypeEnum.INCIDENT_REPORT.value,
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+        earmarked_stocks = self.earmarked_stocks
+        if end_date:
+            earmarked_stocks = earmarked_stocks.filter(created_at__lte=end_date)
+
+        for stock in earmarked_stocks:
+            if stock.earmarked_stock_type == EarmarkedStock.EarmarkedStockChoices.CREATED:
+                action = "Earmarked created"
+                if stock.campaign:
+                    action += f" for {stock.campaign.obr_name}"
+                    if stock.round:
+                        action += f" Round {stock.round.number}"
+
+                base_result = {
+                    "date": stock.created_at.date(),
+                    "action": action,
+                    "vials_in": None,
+                    "doses_in": None,
+                    "vials_out": stock.vials_earmarked,
+                    "doses_out": stock.doses_earmarked,
+                    "type": "earmarked_stock__created",
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+            elif stock.earmarked_stock_type == EarmarkedStock.EarmarkedStockChoices.RETURNED:
+                action = "Earmarked returned"
+                if stock.campaign:
+                    action += f" for {stock.campaign.obr_name}"
+                    if stock.round:
+                        action += f" Round {stock.round.number}"
+                base_result = {
+                    "date": stock.created_at.date(),
+                    "action": action,
+                    "vials_in": stock.vials_earmarked,
+                    "doses_in": stock.doses_earmarked,
+                    "vials_out": None,
+                    "doses_out": None,
+                    "type": "earmarked_stock__returned",
+                }
+
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+        return results
+
+    def get_list_of_used_vials(self, end_date=None, expanded=False):
+        # Used vials are those related to formA outgoing movements. Vials with e.g expired date become unusable, but have not been used
+        outgoing_movements = OutgoingStockMovement.objects.filter(vaccine_stock=self.vaccine_stock)
+        additional_fields = {
+            "id": self.vaccine_stock.id,
+            "country_name": self.vaccine_stock.country.name,
+            "country_id": self.vaccine_stock.country.id,
+            "vaccine_type": self.vaccine_stock.vaccine,
+            "vials_type": "usable",
+        }
+        if end_date:
+            outgoing_movements = outgoing_movements.filter(report_date__lte=end_date)
+        results = []
+        for movement in outgoing_movements:
+            if movement.usable_vials_used > 0:
+                if movement.earmarked_stocks.count() > 0:
+                    earmarked_stock_vials = (
+                        movement.earmarked_stocks.aggregate(total=Sum("vials_earmarked"))["total"] or 0
+                    )
+                    desc_text = f"Form A - Vials Used ({earmarked_stock_vials} vials from Earmarked)"
+
+                else:
+                    desc_text = "Form A - Vials Used"
+                base_result = {
+                    "date": movement.report_date,
+                    "action": desc_text,
+                    "vials_out": None,
+                    "doses_out": None,
+                    "vials_in": movement.usable_vials_used or 0,
+                    "doses_in": (movement.usable_vials_used or 0) * self.get_doses_per_vial(),
+                    "type": MovementTypeEnum.OUTGOING_STOCK_MOVEMENT.value,
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+        return results
+
+    def get_list_of_unusable_vials(self, end_date=None, expanded=False):
+        # First get the used vials
+        results = self.get_list_of_used_vials(end_date, expanded=expanded)
+        additional_fields = {
+            "id": self.vaccine_stock.id,
+            "country_name": self.vaccine_stock.country.name,
+            "country_id": self.vaccine_stock.country.id,
+            "vaccine_type": self.vaccine_stock.vaccine,
+            "vials_type": "usable",
+        }
+        # Get all IncidentReports and Destruction reports for the VaccineStock
+        incident_reports = IncidentReport.objects.filter(vaccine_stock=self.vaccine_stock)
+
+        destruction_reports = DestructionReport.objects.filter(vaccine_stock=self.vaccine_stock).order_by(
+            "destruction_report_date"
+        )
+        if end_date:
+            incident_reports = incident_reports.filter(date_of_incident_report__lte=end_date)
+            destruction_reports = destruction_reports.filter(destruction_report_date__lte=end_date)
+
+        for report in destruction_reports:
+            base_result = {
+                "date": report.destruction_report_date,
+                "action": (f"{report.action}" if len(report.action) > 0 else "Destruction report"),
+                "vials_in": None,
+                "doses_in": None,
+                "vials_out": report.unusable_vials_destroyed or 0,
+                "doses_out": (report.unusable_vials_destroyed or 0) * self.get_doses_per_vial(),
+                "type": MovementTypeEnum.DESTRUCTION_REPORT.value,
+            }
+            if not expanded:
+                results.append(base_result)
+            else:
+                results.append({**base_result, **additional_fields})
+
+        # Add unusable vials from IncidentReports
+        for report in incident_reports:
+            if report.unusable_vials > 0 and (
+                report.stock_correction == IncidentReport.StockCorrectionChoices.PHYSICAL_INVENTORY_ADD
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.VACCINE_EXPIRED
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.VVM_REACHED_DISCARD_POINT
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.UNREADABLE_LABEL
+                or report.stock_correction == IncidentReport.StockCorrectionChoices.BROKEN
+            ):
+                base_result = {
+                    "date": report.date_of_incident_report,
+                    "action": report.stock_correction,  # for every field FOO that has choices set, the object will have a get_FOO_display() method
+                    "vials_in": report.unusable_vials or 0,
+                    "doses_in": (report.unusable_vials or 0) * self.get_doses_per_vial(),
+                    "vials_out": None,
+                    "doses_out": None,
+                    "type": MovementTypeEnum.INCIDENT_REPORT.value,
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+            if report.unusable_vials > 0 and (
+                report.stock_correction == IncidentReport.StockCorrectionChoices.PHYSICAL_INVENTORY_REMOVE
+            ):
+                base_result = {
+                    "date": report.date_of_incident_report,
+                    "action": report.stock_correction,  # for every field FOO that has choices set, the object will have a get_FOO_display() method
+                    "vials_in": None,
+                    "doses_in": None,
+                    "vials_out": report.unusable_vials or 0,
+                    "doses_out": (report.unusable_vials or 0) * self.get_doses_per_vial(),
+                    "type": MovementTypeEnum.INCIDENT_REPORT.value,
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+        # Add earmarked stock movements of type USED
+        earmarked_stocks = self.earmarked_stocks.filter(
+            vaccine_stock=self.vaccine_stock,
+            earmarked_stock_type=EarmarkedStock.EarmarkedStockChoices.USED,
+        )
+
+        if end_date:
+            earmarked_stocks = earmarked_stocks.filter(created_at__date__lte=end_date)
+
+        for stock in earmarked_stocks:
+            if (
+                stock.earmarked_stock_type == EarmarkedStock.EarmarkedStockChoices.USED and stock.form_a is None
+            ):  # if FormA is not None, it's accounted by the FormA, no need to repeat
+                base_result = {
+                    "date": stock.created_at.date(),
+                    "action": f"Earmarked stock used for {stock.campaign.obr_name} Round {stock.round.number}",
+                    "vials_in": stock.vials_earmarked,
+                    "doses_in": stock.doses_earmarked,
+                    "vials_out": None,
+                    "doses_out": None,
+                    "type": "earmarked_stock__used",
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+        return results
+
+    def get_list_of_earmarked(self, end_date=None, expanded=False):
+        earmarked_movements = self.earmarked_stocks
+        additional_fields = {
+            "id": self.vaccine_stock.id,
+            "country_name": self.vaccine_stock.country.name,
+            "country_id": self.vaccine_stock.country.id,
+            "vaccine_type": self.vaccine_stock.vaccine,
+            "vials_type": "usable",
+        }
+        if end_date:
+            earmarked_movements = earmarked_movements.filter(created_at__lte=end_date)
+
+        results = []
+        for movement in earmarked_movements:
+            movement_type = movement.earmarked_stock_type
+            if (
+                movement_type == EarmarkedStock.EarmarkedStockChoices.USED
+                or movement_type == EarmarkedStock.EarmarkedStockChoices.RETURNED
+            ):
+                if movement.form_a is not None:
+                    action_text = f"Earmarked stock used for FormA ({movement.form_a})"
+                else:
+                    action_text = "Earmarked stock used"
+                    if movement.campaign:
+                        action_text += f" for {movement.campaign.obr_name}"
+                        if movement.round:
+                            action_text += f" Round {movement.round.number}"
+
+                base_result = {
+                    "date": movement.created_at.date(),
+                    "action": action_text,
+                    "vials_out": movement.vials_earmarked,
+                    "doses_out": movement.doses_earmarked,
+                    "vials_in": None,
+                    "doses_in": None,
+                    "type": f"earmarked_stock__{movement_type}",
+                }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+            else:
+                action_text = "Earmarked stock reserved"
+                if movement.campaign:
+                    action_text += f" for {movement.campaign.obr_name}"
+                    if movement.round:
+                        action_text += f" Round {movement.round.number}"
+                elif movement.temporary_campaign_name:
+                    action_text += f" for ({movement.temporary_campaign_name})"
+
+                    base_result = {
+                        "date": movement.created_at.date(),
+                        "action": action_text,
+                        "vials_in": movement.vials_earmarked,
+                        "doses_in": movement.doses_earmarked,
+                        "vials_out": None,
+                        "doses_out": None,
+                        "type": f"earmarked_stock__{movement_type}",
+                    }
+                if not expanded:
+                    results.append(base_result)
+                else:
+                    results.append({**base_result, **additional_fields})
+
+        return results
