+++ plugins/polio/tasks/archive_vaccine_stock_for_rounds.py
@@ -6,9 +6,8 @@ from django.db.models import Q
 from django.db.utils import IntegrityError
 
 from beanstalk_worker import task_decorator
-from plugins.polio.api.vaccines.stock_management import VaccineStockCalculator
 from plugins.polio.models import Round, VaccineStock, VaccineStockHistory
-from plugins.polio.models.base import VACCINES
+from plugins.polio.models.base import VACCINES, VaccineStockCalculator
 
 
 logger = logging.getLogger(__name__)
